refactor(core): use usize for block_ids, validate at RPC boundary
Block ids are array indices, so the engine now speaks usize for them
instead of i32. The i32->usize conversion (which rejects negatives)
happens once in service.rs for the load and save handlers, mirroring
the existing scalar validation pattern, rather than per-block deep
inside the layers x blocks loop in lib.rs and offload.rs.

device_id stays i32 since it is a CUDA device handle, not an index.
The PyO3 gRPC client keeps Vec<i32> since that is the protobuf wire type.

Verified with cuda-13 feature: pegaflow-core and pegaflow-server unit
tests pass (save_validation, prefetch_lease included). Skipped the hook's
full GPU suite; http_cleanup_hang_repro crashes pre-existing on clean
master here (torch CUDA/GIL interpreter conflict), unrelated to this change.

BREAKING CHANGE: batch_load_kv_blocks_multi_layer(_inproc) and
LayerSave.block_ids now take usize instead of i32.

+++ pegaflow-server/examples/p2p_bench.rs
@@ -332,7 +332,7 @@ async fn run_holder(cli: &Cli, shape: &Shape, pool_bytes: usize) {
     });
 
     let ranks = register_ranks(&engine, shape);
-    let block_ids: Vec<i32> = (0..shape.blocks as i32).collect();
+    let block_ids: Vec<usize> = (0..shape.blocks).collect();
     let layer_names = shape.layer_names();
 
     let mut image = vec![0u8; shape.rank_bytes()];
@@ -473,7 +473,7 @@ async fn verify_set(
         PrefetchStatus::Loading => panic!("verify: set {set} still loading"),
     };
     assert_eq!(blocks.len(), shape.blocks, "verify: incomplete set {set}");
-    let block_ids: Vec<i32> = (0..shape.blocks as i32).collect();
+    let block_ids: Vec<usize> = (0..shape.blocks).collect();
     let layer_names = shape.layer_names();
     let layer_name_refs: Vec<&str> = layer_names.iter().map(|s| s.as_str()).collect();
